Move the overridden `size` regression test into its own file
pretty-format-overflow.test.ts is specific to the deep-nesting overflow
regression; keep the overridden-size coverage separate.
