+++ test/js/bun/test/pretty-format-overflow.test.ts
@@ -51,53 +51,3 @@ test("deep nesting", () => {
     expect(stderr).toContain("expect(received).toEqual(expected)");
   }, 30000);
 });
-
-describe("pretty_format should handle collections with an overridden `size` property", () => {
-  test("non-numeric `size` on (Weak)Set/(Weak)Map still produces a toEqual diff", async () => {
-    await using proc = Bun.spawn({
-      cmd: [
-        bunExe(),
-        "-e",
-        `
-const values = [];
-{
-  const weakSet = new WeakSet();
-  weakSet.size = BigUint64Array;
-  values.push(weakSet);
-}
-{
-  const weakMap = new WeakMap();
-  weakMap.size = "not a number";
-  values.push(weakMap);
-}
-{
-  const set = new Set([1]);
-  Object.defineProperty(set, "size", { value: {} });
-  values.push(set);
-}
-{
-  const map = new Map([[1, 2]]);
-  Object.defineProperty(map, "size", { value: BigUint64Array });
-  values.push(map);
-}
-for (const value of values) {
-  try {
-    Bun.jest().expect(BigUint64Array).toEqual(value);
-    console.log("DID NOT THROW");
-  } catch (e) {
-    console.log(e.message.includes("expect(received).toEqual(expected)") ? "DIFF OK" : "UNEXPECTED: " + e.message);
-  }
-}
-`,
-      ],
-      env: bunEnv,
-      stderr: "pipe",
-      stdout: "pipe",
-    });
-
-    const [stdout, , exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-
-    expect(stdout.trim().split("\n")).toEqual(["DIFF OK", "DIFF OK", "DIFF OK", "DIFF OK"]);
-    expect(exitCode).toBe(0);
-  });
-});
